# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TwinOps AI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/simulation/inject/{scenario_id}": {
      "post": {
        "tags": [
          "twin"
        ],
        "summary": "Inject",
        "operationId": "inject_api_v1_simulation_inject__scenario_id__post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incident"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/replay": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Replay",
        "operationId": "get_replay_api_v1_incidents__incident_id__replay_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/narrative": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Narrative",
        "operationId": "get_narrative_api_v1_incidents__incident_id__narrative_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Narrative"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthState": {
        "properties": {
          "auth_enabled": {
            "type": "boolean",
            "title": "Auth Enabled"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          }
        },
        "type": "object",
        "required": [
          "auth_enabled",
          "user"
        ],
        "title": "AuthState"
      },
...
      "GraphSnapshot": {
        "properties": {
          "topology_hash": {
            "type": "string",
            "title": "Topology Hash"
          },
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "active_scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Scenario Id"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TwinNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/TwinEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "health": {
            "items": {
              "$ref": "#/components/schemas/NodeHealth"
            },
            "type": "array",
            "title": "Health"
          }
        },
        "type": "object",
        "required": [
          "topology_hash",
          "tick",
          "active_scenario_id",
          "nodes",
          "edges",
          "health"
        ],
        "title": "GraphSnapshot",
        "description": "Full state a client needs to render the twin (REST snapshot, \u00a74 pattern)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Incident": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "detected",
              "investigating",
              "identified",
              "resolved"
            ],
            "title": "Status"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Severity"
          },
          "started_tick": {
            "type": "integer",
            "title": "Started Tick"
          },
          "resolved_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Tick"
          },
          "affected_node_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Node Ids"
          },
          "root_cause": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RootCause"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "severity",
          "started_tick",
          "resolved_tick",
          "affected_node_ids",
          "root_cause"
        ],
        "title": "Incident"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "Metrics": {
        "properties": {
          "cpu": {
            "type": "number",
            "title": "Cpu"
          },
          "memory": {
            "type": "number",
            "title": "Memory"
          },
          "latency_p95": {
            "type": "integer",
            "title": "Latency P95"
          },
          "error_rate": {
            "type": "number",
            "title": "Error Rate"
          }
        },
        "type": "object",
        "required": [
          "cpu",
          "memory",
          "latency_p95",
          "error_rate"
        ],
        "title": "Metrics"
      },
      "Narrative": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "source": {
            "type": "string",
            "enum": [
              "llm",
              "deterministic"
            ],
            "title": "Source"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          }
        },
        "type": "object",
        "required": [
          "text",
          "source",
          "provider"
        ],
        "title": "Narrative"
      },
      "NodeHealth": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "critical"
            ],
            "title": "Status"
          },
          "metrics": {
            "$ref": "#/components/schemas/Metrics"
          }
        },
        "type": "object",
        "required": [
          "id",
          "score",
          "status",
          "metrics"
        ],
        "title": "NodeHealth"
      },
...
      "ReplayFrame": {
        "properties": {
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "health": {
            "items": {
              "$ref": "#/components/schemas/NodeHealth"
            },
            "type": "array",
            "title": "Health"
          }
        },
        "type": "object",
        "required": [
          "tick",
          "health"
        ],
        "title": "ReplayFrame"
      },
      "ReplayResponse": {
        "properties": {
          "incident_id": {
            "type": "string",
            "title": "Incident Id"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/ReplayFrame"
            },
            "type": "array",
            "title": "Frames"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "origin",
          "frames"
        ],
        "title": "ReplayResponse"
      },
      "RootCause": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "estimated_recovery": {
            "type": "string",
            "title": "Estimated Recovery"
          },
          "runbook_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runbook Id"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "label",
          "confidence",
          "summary",
          "evidence",
          "recommended_actions",
          "estimated_recovery",
          "runbook_id"
        ],
        "title": "RootCause",
        "description": "Deterministic root-cause verdict inferred from topology + health.\nConfidence is computed from evidence weights, never model vibes (\u00a70.3)."
      },
...
      "TwinEdge": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "source",
          "target"
        ],
        "title": "TwinEdge"
      },
      "TwinNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "load-balancer",
              "gateway",
              "service",
              "database",
              "queue",
              "cache",
              "cloud"
            ],
            "title": "Kind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "tier": {
            "type": "integer",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "label",
          "tier"
        ],
        "title": "TwinNode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}